Refactor: switched isLoggedIn selector to register/selector
