--- src/redux/auth/selectors.js
@@ -1,1 +1,0 @@
-export const selectIsLoggedIn = (state) => state.auth.isLoggedIn;
